requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
django-bootstrap-datepicker==1.2.2
django-crispy-forms==1.6.1
django-secure==1.0.1
django-sslserver==0.19
gunicorn==19.6.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
PyJWT==1.4.2
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
reportlab==3.3.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
whitenoise==3.3.0
XlsxWriter==0.9.6
